refactor(web): componentsを役割別カテゴリで整理
ui配下の独自コンポーネントをデザインシステムの分類に倣い
layout / navigation / typography / feedback の役割別に再配置する。
階層が深くなるためフォルダ外参照は@/clientエイリアスへ統一。
