+++ application/packages/web/src/client/components/ui/navigation/nav-menu/index.test.ts
@@ -1,8 +1,8 @@
 import type { ReactNode } from 'react'
 import { Children, isValidElement } from 'react'
 import { describe, expect, it } from 'vitest'
-import { SheetClose } from '../../shadcn/sheet'
-import type { MenuItem } from '../sidebar'
+import { SheetClose } from '@/client/components/shadcn/sheet'
+import type { MenuItem } from '@/client/components/ui/layout/sidebar'
 import NavMenu from './index'
 
 const menuItems: MenuItem[] = [
