+++ application/packages/web/src/client/components/ui/typography/clip-text/index.tsx
