+++ application/packages/web/src/client/features/diary/diary-read-list-section.tsx
@@ -1,6 +1,6 @@
 import { toJaTimeString } from '@trend-diary/common/locale/date'
 import type { ReactNode } from 'react'
-import { AnchorLink } from '@/client/components/ui/link'
+import { AnchorLink } from '@/client/components/ui/navigation/link'
 import type { ReadItem } from '@/client/features/diary/types'
 import { toSafeExternalPath } from '@/client/lib/url'
 import MediaIcon from '@/client/routes/trends._index/components/media-icon'
