+++ application/packages/web/src/client/lib/url.ts
@@ -1,4 +1,4 @@
-import type { ExternalPath } from '@/client/components/ui/link'
+import type { ExternalPath } from '@/client/components/ui/navigation/link'
 
 export function toSafeExternalPath(url: string): ExternalPath | null {
   try {
